+++ packages/tools/utils/src/object-boolean.spec.js
@@ -1,6 +1,11 @@
 import {strict as assert} from "assert";
 
-import {hasObjSize1, isObjEmpty, isObjNotEmpty} from "./object-boolean";
+import {
+	hasObjSize1,
+	hasSomeNullValues,
+	isObjEmpty,
+	isObjNotEmpty
+} from "./object-boolean";
 
 describe("Object -> Boolean", function() {
 	describe("hasObjSize1", function() {
@@ -12,6 +17,18 @@ describe("Object -> Boolean", function() {
 		});
 		it("should return `false` if the object is longer than 1", function() {
 			assert.deepStrictEqual(hasObjSize1({a: 1, b: 2}), false);
+		});
+	});
+	describe("hasSomeNullValues", function() {
+		it("should return false is the provided object has no properties being `null`", function() {
+			assert.deepStrictEqual(hasSomeNullValues({a: 1}), false);
+			assert.deepStrictEqual(hasSomeNullValues({a: 1, b: undefined}), false);
+		});
+		it("should return true is some of the provided object properties are `null`", function() {
+			assert.deepStrictEqual(
+				hasSomeNullValues({a: 1, b: undefined, c: null}),
+				true
+			);
 		});
 	});
 	describe("isObjEmpty", function() {