+++ CHANGELOG.md
@@ -29,6 +29,14 @@
 
 ## `@svizzle/utils` v0.8.0 (next)
 
+### (String -> String) -> (Object -> Object)
+
+- add `renameKeysWith`
+
+### Any -> (Any -> Boolean)
+
+- add `isEqualTo`
+
 ### Array -> (Array -> Array)
 
 - add `pluckKeys`
@@ -41,9 +49,9 @@
 
 - Fix `inclusiveRange` returning an empty array when the extent has 2 zeroes
 
-### (String -> String) -> (Object -> Object)
+### Object -> Boolean
 
-- add `renameKeysWith`
+- add `hasSomeNullValues`
 
 ## 20200411_2